fixing compiler warning
SVN-Revision: 29761

+++ contrib/brl/bseg/boxm/ocl/view/boxm_update_bit_tableau.cxx
@@ -186,7 +186,7 @@ bool boxm_update_bit_tableau::update_model()
   count_++; curr_count_++;
 
   //make sure you get a valid frame...
-  if(curr_frame_ >= cam_files_.size())
+  if(curr_frame_ >= (int)cam_files_.size())
     curr_frame_ = 0;
   curr_frame_=rand.lrand32(0,cam_files_.size()-1);
